% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy2_and}
\alias{fuzzy2_and}
\title{Funktion fuzzy2_and(pf1,pf2)}
\usage{
fuzzy2_and(pf1, pf2)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}

\item{pf2}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Ein Vektor mit Fuzzywerten
}
\description{
berechnet das fuzzylogischen UND
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1.
Grundlage für die fuzzy2_...() ist hier multiplikative Verknüpfung für UND
Dies ist eine strenges Fuzzy-UND im Vergleich zu der fuzzy_and() als  Minimum
}
