% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy2_or}
\alias{fuzzy2_or}
\title{Funktion fuzzy2_or(pf1,pf2)
berechnet das fuzzylogische ODER mit dem Maximum von pf1 und pf2
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1.
Grundlage für die fuzzy2_or() ist hier Negation der multiplikativen Verknüpfung für UND
Dies liefert ein schwächeres Fuzzy-ODER im Vergleich zu der fuzzy_or() als Maximum}
\usage{
fuzzy2_or(pf1, pf2)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}

\item{pf2}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Zahl oder Vektor mit Fuzzywerten
}
\description{
Funktion fuzzy2_or(pf1,pf2)
berechnet das fuzzylogische ODER mit dem Maximum von pf1 und pf2
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1.
Grundlage für die fuzzy2_or() ist hier Negation der multiplikativen Verknüpfung für UND
Dies liefert ein schwächeres Fuzzy-ODER im Vergleich zu der fuzzy_or() als Maximum
}
