% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy_and}
\alias{fuzzy_and}
\title{Funktion fuzzy_and(pf1,pf2)}
\usage{
fuzzy_and(pf1, pf2)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}

\item{pf2}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Ein Vektor mit Fuzzywerten
}
\description{
berechnet das fuzzylogische UND  mit dem Minimum von pf1 und pf2
von zwei Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1.
}
