% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{fuzzy_not}
\alias{fuzzy_not}
\title{Funktion fuzzy_not(pf1)}
\usage{
fuzzy_not(pf1)
}
\arguments{
\item{pf1}{Zahl oder Vektor mit Fuzzywerten}
}
\value{
Zahl oder Vektor mit Fuzzywerten
}
\description{
berechnet das fuzzylogische NICHT mit 1-pf1
von einem Vektoren pf1 und pf2 von Fuzzywerte. Die Komponenten der Vektoren
sind reelle Zahlen zwischen 0 und 1.
}
