% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{koaktivitaetsmatrix}
\alias{koaktivitaetsmatrix}
\title{Funktion koaktivitaetsmatrix}
\usage{
koaktivitaetsmatrix(pVecX, pVecY)
}
\arguments{
\item{pVecX}{Ein Vektor}

\item{pVecY}{Ein Vektor}
}
\value{
Ein Matrix
}
\description{
Funktion koaktivitaetsmatrix
}
\examples{
koaktivitaetsmatrix(c(1, 2, 3), c(4, 5, 6))
koaktivitaetmatrix()
x <- c(1,0,1)
y <- c(0,1)
koaktivitaetsmatrix(x,y)
## Matrix hat die Gestalt
##       [,1] [,2] [,3]
## [1,]    0    0    0
## [2,]    1    0    1

}
