% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{loesung_ausgeben}
\alias{loesung_ausgeben}
\title{Funktion loesung_ausgeben(pHilfe,pName,pSolution)}
\usage{
loesung_ausgeben(hilfe, name, solution)
}
\arguments{
\item{hilfe}{String/Zeichenkette mit einer Hilfe}

\item{name}{Der Name}

\item{solution}{Die Lösung}
}
\value{
Die Hilfe mit ersetztem Name und ersetzter Lösung
}
\description{
Die Funktion erhält eine Hilfe als Zeichenkette. Die Zeichenkette beinhaltet
Textstellen mit "NAME" und "SOLUTION".
Die Textstelle "NAME" wird durch den Inhalt des Parameters "pName" ersetzt.
Die Textstelle "SOLUTION" wird durch den Inhalt des Parameters "pSolution" ersetzt.
}
\examples{
hilfe <- "Hinweis für NAME : Die Lösung ist SOLUTION"
loesung_ausgeben(hilfe, "Hans", "zerlege 8 in 3+5 und rechne 7+8=(7+3)+5 ")

}
