% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr4education.R
\name{notengrenzen_calc}
\alias{notengrenzen_calc}
\title{Funktion notengrenzen_calc}
\usage{
notengrenzen_calc(pPunkteMax, pBestGrenze, pBestIndex, pNotenSkala)
}
\arguments{
\item{pPunkteMax}{maximale Anzahl der Punkte in der Klausur}

\item{pBestGrenze}{Bestehensgrenze, ab der die Klausur bestanden ist}

\item{pBestIndex}{Bestehensindex, ist der Index in der Notenskala, bei dem die Klausur gerade noch bestanden ist}

\item{pNotenSkala}{Notenskala als Vektor der bestandenen Werte}
}
\value{
ein Vektor, der jeweils die Notengrenzen für die jeweiligen Noten angibt (von guten zu schlechteren Noten)
}
\description{
Mit pPunkteMax als Parameter wird die maximale Anzahl der Punkte in der Klausur festgelegt.
Jede Klausur hat eine Bestehensgrenze, mit der man gerade noch bestanden hat. Diese wird durch
den Parameter pBestGrenze angegeben. Gleichzeitig bekommt die Funktion auch eine Notenskala, die als
Grundlage für die später Bewertung bzw. Notenzuordnung verwendet wird.
Ein Beispiel für eine Notenskala ist z.B. der Vektor mit Noten
c("sehr gut","gut","befriedigend",ausreichend","mangelhaft","ungenügend")
Dazu wird ein Bestehensindex mit pBestIndex als Parameter übergeben. Der Bestehensindex ist im
obigen Beispiel  4 und verweist damit auf die Note "ausreichend" mit der man gerade noch bestanden hat.
Notenskalen können mit dem zugehörigen Bestehensindex auf wie folgt aussehen.
notenskala <- c("1+","1","1-","2+","2","2-","3+","3","3-","4+","4","4-","5+","5","5-","6")
bestindex <- 11 verweist als Index für Note "4" und mit "4-" hat man nicht mehr bestanden.
notenskala <- c("1+","1","1-","2+","2","2-","3+","3","3-","4+","4","4-","5+","5","5-","6")
bestindex <- 12 verweist als Index für Note "4-"  mit der nun noch bestanden hat.
Notenskalen können auch Zahlenwerte enthalten. Die kann von Vorteil sein, wenn man später zugeordnet Noten
auf ganze Noten runden möchte.
notenskala <- c(0.7,1.0, 1.3, 1.7, 2.0, 2.3, 3.3, 3.7, 4.0, 4.3, 4.7, 5.0, 5.3, 6.0)
Die Funktion notengrenzen_calc() wird zusammen mit der Funktion noten_zuordnen() verwendet, die die Notengrenze
als Eingabe benötigt, um die Noten zuordnen zu können.
}
\examples{
notenskala <- c("sehr gut", "gut", "befriedigend", "ausreichend", "mangelhaft", "ungenügend")
notengrenzen_calc(100, 50, 4, notenskala)
## Ausgabe: [1] 100.0  87.5  75.0  62.5  50.0  25.0
## Bestanden haette man mit "ausreichend". Der Bestehensindex ist also in diesem Fall 4

}
